changed relevant datatypes to proper work with 64bit architecture

+++ userspace/libc/include/pthread.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 
 //pthread typedefs
-typedef unsigned int pthread_t;
+typedef long unsigned int pthread_t;
 typedef unsigned int pthread_attr_t;
 
 //pthread mutex typedefs
