+++ userspace/libc/include/wait.h
@@ -9,7 +9,7 @@ extern "C" {
 //pid typedefs
 #ifndef PID_T_DEFINED
 #define PID_T_DEFINED
-typedef int pid_t;
+typedef long int pid_t;
 #endif // PID_T_DEFINED
 
 extern pid_t waitpid(pid_t pid, int *status, int options);
